Skips parsing for the 50 most frequent fields in OSM (on 20 june 2018) by storing their trees
The file "frequent_fields.py" now stores parsing trees of the 50 most frequent fields in OpenStreetMap.
So, if the module has to parse one of them, it can just take its tree in the dict, avoiding the need to parse it.
These fields represent about 30% of all the fields in OSM, and their use make the "__init__" method about 16 times faster.
So, in about 30% of cases, the parsing will be about 16 times faster (because the field won't be parsed).