Fixes a bug in the 'get_current_rule()' method and adds unit tests for it
When a field had different rules for a date (ex: "Mo-Fr 10:00-20:00; Dec off"), only the first tested rule (with 'is_included()') was considered, so 'is_open()' returned True for a Monday in December (it took the first rule, matching for any day from Monday to Friday, instead of the other matching for December).
This commit fixes this bug and adds unit tests to ensure it won't occur again.